requirements.txt
asgiref==3.3.1
beautifulsoup4==4.9.3
boto3==1.14.2
botocore==1.17.2
certifi==2020.4.5.2
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.2.16
django-appconf==1.0.4
django-cors-headers==3.3.0
django-elasticsearch-dsl==7.1.1
django-filter==2.3.0
django-nine==0.2.3
django-parler==1.9.2
django-rest-swagger==2.2.0
django-storages==1.10.1
djangorestframework==3.11.0
docutils==0.15.2
drf-yasg==1.20.0
elasticsearch==7.7.1
elasticsearch-dsl==7.2.1
gunicorn==20.0.4
idna==2.5
inflection==0.5.1
install==1.3.4
itypes==1.2.0
Jinja2==2.11.2
jmespath==0.10.0
MarkupSafe==1.1.1
openapi-codec==1.3.2
packaging==20.4
Pillow==7.1.2
psycopg2==2.8.4
pyparsing==2.4.7
python-dateutil==2.8.1
python-decouple==3.3
pytils==0.3
pytz==2020.1
requests==2.21.0
requests-aws4auth==1.0
requests2==2.16.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
s3transfer==0.3.3
simplejson==3.17.2
six==1.14.0
soupsieve==2.0.1
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.24.3
whitenoise==5.2.0
